Add dependencies to same version of libnma
nm-applet and nm-connection-editor RPMs already depend on libnma
symbols. Make the dependency stricter by specifying the exact version
of the library. This makes it to easier to debug problems reported by
users as we don't have to test different combinations of
binaries/libraries.

Also add the %{?_isa} suffix to dependencies.

https://bugzilla.redhat.com/show_bug.cgi?id=1679251